+++ TestingAwsUpload/SymbolDetail.swift
@@ -25,30 +25,3 @@ struct SymbolDetail: View {
         .padding()
     }
 }
-
-//struct SymbolDetail: View {
-//    var symbol: Symbol
-//
-//
-//    var body: some View {
-//        VStack {
-//            Text(symbol.name)
-//                .font(.system(.largeTitle, design: .rounded))
-//
-//            Image(systemName: symbol.name)
-//                .resizable()
-//                .scaledToFit()
-//                .symbolRenderingMode(.hierarchical)
-//                .foregroundColor(.accentColor)
-//        }
-//        .padding()
-//    }
-//}
-
-
-struct Details_Previews: PreviewProvider {
-    static var previews: some View {
-        SymbolDetail(symbol: CoverImage(name: "New Cutie", cover: Image("hb1")))
-//        SymbolDetail(symbol: Symbol(name: "magnifyingglass"))
-    }
-}
